We aren't using this anywhere
Also the logic doesn't make whole lot of sense. Surely if we're going to count the size it should be in bytes? If we did actually want the number if items surely we'd also want the number of other things like queries.

Since it's not used I figure it's best to remove for noe