Add ListChangesByStatus for https://github.com/overmindtech/ovm-cli/issues/32
Adding a more specific `...ByStatusAndTicketLink` would reduce the amount
of code in the cli by a little bit at the cost of making a otherwise
useless function here. We expect that there will be only few changes in
DEFINING at any point in time, keeping the overall performance impact in
check.