Move to deadlines (pt1)
Deadlines (point-in-time) are more robust and easier to handle than
timeouts (relative) as they propagate without needing any processing
and still work in the face of variable transmission delays and other
intermediate steps.

To avoid breaking the protocol, this is a add-only change. Once everyone
has switched over to using deadlines, we can remove the timeout fields.