Endpoints for HCP Terraform integration
To enable/configure the HCP Terraform integration, we need to create an
API Key to hold the user's authorization. To capture the authorization
this will go through the same redirect lap as `apikeys.CreateAPIKey` does.
To facilitate code-reuse, `CreateHcpConfig` re-uses the messages from the
former and will pass through the request internally.