Improve docs for terraform mappings & remove scope
This improves the docs and removes the scope mapping. The scope mapping was previously used in conjunction with the `overmind_mappings` Terraform output that allowed the mapping to select the correct scope. This has since been fixed with a much smarter matching system which means that this isn't needed. The docs were removed a while ago and we're now going to remove all of this functionality